Conditional packages no return an installed path of `None` if not local.
I.e. if the package is on the system a install path of `None` is returned, packages can then decide how to link to the conditional package based on this.
The snoplus/rat package now uses this to check whether to add environment information for the conditional packages.
Also fixed snoplus/rat is_installed library checks.